#56 Fix test execution and failing test
- Move tests to separate test directory
  - Adapt mocha config files
- Extract helper methods to separate files
- Remove misplaced `only` annotation in model-service integration test file

Contributed on behalf of STMicroelectronics
